add `handlePayRequest`
